RPC refactoring
