+++ rai/rpc/rpc_secure.cpp
